+++ NEWS.md
@@ -4,7 +4,7 @@
   boundary elements of `<grates>` objects. Not guaranteed to end up on CRAN
   (and may be renamed if they do).
 
-* Experimental (untested) function `%during%` for testing whether a scalar
+* Experimental function `%during%` for testing whether a scalar
   date is contained within the range (inclusive) of a <grates> object. Not
   guaranteed to end up on CRAN (and may be renamed if they do).