+++ NEWS.md
@@ -1,3 +1,5 @@
+# grates (development version)
+
 # grates 1.2.0
 
 * New functions `date_start()` and `date_end()` for accessing the